% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots_cat.R
\name{bar_cat}
\alias{bar_cat}
\title{A barplot showing either the absolute or relative appearances of the levels of the input factor. Moreover a visual categories check can be performed by supplying a character vector of allowed categories.}
\usage{
bar_cat(
  x_factor,
  var_name = "",
  categories = NULL,
  length_original = NULL,
  color1 = "darkviolet",
  color2 = "yellow2",
  alpha = 0.8,
  ttl = "",
  rel = FALSE,
  flip_axis = FALSE
)
}
\arguments{
\item{x_factor}{A factor vector.}

\item{var_name}{The variable name to be displayed under the boxplot. Must be given as character.}

\item{categories}{A character vector of categories. Categories that are not included in this vector will be highlighted on the plot.}

\item{length_original}{If the factor vector was already filtered one can supply an original length to get accurate results in the relatvie version of the barplot.}

\item{color1}{The primary filling color of the barplot. Must be given as character.}

\item{color2}{The secondary filling color of the barplot indicating categories that were not included in the supplied ones.}

\item{alpha}{A numeric value between 0 and 1 to adjust the transparancy of the barplot.}

\item{ttl}{An optional title for the plot. Must be given as charcater.}

\item{rel}{Logical value indicating whether the count or the density should be shown on the y axis. (default: count)}

\item{flip_axis}{Logical value indicating whether categories should be shown on the x or y axis. (default: x axis) Especially for long category names one should consider the axis flip.}
}
\value{
a ggplot2 object
}
\description{
A barplot showing either the absolute or relative appearances of the levels of the input factor. Moreover a visual categories check can be performed by supplying a character vector of allowed categories.
}
\examples{
bar_cat(factor(c(1:10,4,4,2,7,7,7,7)))
}
\author{
Emanuel Sommer
}
