% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_vector.R
\name{classify_vector}
\alias{classify_vector}
\title{Determine whether the vector will be classified as: continuous, categorical, datetime or date}
\usage{
classify_vector(x, cat_thres = 30)
}
\arguments{
\item{x}{a single vector (the allowed classes are: numeric, integer, factor, logical, character, POSIXct/ POSIXlt for datetimes and Date for dates.)}

\item{cat_thres}{non negative integer value to set the number of unique values a numeric or integer field needs to have to be classified as continuous. (default is set to 30)}
}
\value{
a character string with either "cont", "cat", "datetime or "date"
}
\description{
Determine whether the vector will be classified as: continuous, categorical, datetime or date
}
\author{
Emanuel Sommer
}
