% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots_cont.R
\name{cont_boxplot}
\alias{cont_boxplot}
\title{Generate a boxplot with jittered points for a single numeric vector.}
\usage{
cont_boxplot(
  x,
  var_name = "",
  color = "darkviolet",
  size = 1,
  width = 0.35,
  title = ""
)
}
\arguments{
\item{x}{A numeric vector.}

\item{var_name}{The variable name to be displayed under the boxplot. Must be given as character.}

\item{color}{The color of the boxplot. Must be given as character.}

\item{size}{The size of the lines of the boxplot. Must be given as a positive numeric value.}

\item{width}{The width of the boxplot. Must be given as a positive numeric value.}

\item{title}{The title of the plot. Must be given as character.}
}
\value{
a ggplot2 object
}
\description{
Generate a boxplot with jittered points for a single numeric vector.
}
\examples{
cont_boxplot(runif(100))
}
\author{
Emanuel Sommer
}
