% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 13 columns.
}
\usage{
dummy_data
}
\description{
Randomly generated data that allows to play around with the functions of the package. Have fun :)
}
\keyword{datasets}
