% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof_fun.R
\name{extract_rows_score}
\alias{extract_rows_score}
\title{The function is a shortcut to extract the relevant rows of the input data of the lof_fun function.}
\usage{
extract_rows_score(data, lof_list, threshold = 1)
}
\arguments{
\item{data}{The tibble or dataframe used in the call of lof_fun.}

\item{lof_list}{The output list of the lof_fun function.}

\item{threshold}{Greater scores than the threshold will be extracted.}
}
\value{
tibble or dataframe with the relevant rows from data + a column with the LOF-scores
}
\description{
The function is a shortcut to extract the relevant rows of the input data of the lof_fun function.
}
\author{
Emanuel Sommer
}
