% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots_date_datetime.R
\name{hist_datetime}
\alias{hist_datetime}
\title{Generate a histogram of a datetime input vector. Moreover range bounds can be included in the visualiation and a relative or absolute histogram can be selected}
\usage{
hist_datetime(
  x,
  var_name = "",
  color = "darkviolet",
  range_col = "black",
  bin_num = 100,
  ttl = "",
  alpha = 0.8,
  range_min = NULL,
  range_max = NULL,
  rel = FALSE
)
}
\arguments{
\item{x}{A datetime vector (POSIXct).}

\item{var_name}{The variable name to be displayed under the boxplot. Must be given as character.}

\item{color}{The filling color of the histogram. Must be given as character.}

\item{range_col}{The color of the range bounds. Must be given as character}

\item{bin_num}{The number of bins of the histogram. Must be given as a positive integer value.}

\item{ttl}{The title of the plot. Must be given as character.}

\item{alpha}{A numeric value between 0 and 1 to adjust the transparancy of the histogram.}

\item{range_min}{A POSIXct value setting the lower range bound of the data. If the actual data exceeds the bound the bound is going to visually appear on the plot.}

\item{range_max}{A POSIXct value setting the upper range bound of the data. If the actual data exceeds the bound the bound is going to visually appear on the plot.}

\item{rel}{Logical value indicating whether the count or the density should be shown on the y axis. (default: count)}
}
\value{
a ggplot2 object
}
\description{
Generate a histogram of a datetime input vector. Moreover range bounds can be included in the visualiation and a relative or absolute histogram can be selected
}
\author{
Emanuel Sommer
}
