% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof_fun.R
\name{lof_fun}
\alias{lof_fun}
\title{Local outlier factor (LOF)}
\usage{
lof_fun(data, col1 = "darkviolet")
}
\arguments{
\item{data}{A tibble or dataframe with the variables to be checked for outliers as the columns (columns must be named)}

\item{col1}{The color of the boxplot. Must be given as character.}
}
\value{
A list with the elements scores and boxplot.
}
\description{
A function that calculates the Local Outlier Factor scores for a given set of variables with the same length.
Moreover the output list contains a boxplot that visualizes the LOF-scores. The goal is to provide an easy
interface for multivariate outlier detection.
}
\details{
LOF is a nearest neighbors approach where the local densities are compared. The score is basically a division
of the density drawn from the k nearest neighbors divided by the density of the point of interest.
Hence values greater than 1 are considered more likely to be an anomaly. In this function the k is not set
to a specific value. The function uses a tuning vector of values for k, calculates the scores for any of these
k and then averages over all of these scores. This yields the final score that can be found in the output list.
The tuning vector is set up around the square root of the variable length (+/- 10) and no k under 10 will be used
as the inventor of the LOF claims that this would yield to very unrobust scores. For the LOF-Algorithm the
implementation of the dbscan package is used. Moreover the gower distance is used as the metric for the knn algorithm as it can also handle c
categorical input. In this implementation that opts for an easy to use approach, rows with NAs will be omitted. This can result in major problems
if the percentage of missing values is relevant or if the missing values are missing due to a certain pattern. Prior imputing of the missing values is recommended.
}
\author{
Emanuel Sommer
}
