% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lof_fun.R
\name{lof_vis}
\alias{lof_vis}
\title{Visualization for LOF on 2 variables.}
\usage{
lof_vis(data, lof_list, col_low = "deeppink", col_high = "darkviolet")
}
\arguments{
\item{data}{The data that was initially used for the \code{lof_fun} function. (only 2 named columns!)}

\item{lof_list}{The output list of the lof_fun function.}

\item{col_low}{The color for low LOF-scores. Must be given as character.}

\item{col_high}{The color for high LOF-scores. Must be given as character.}
}
\value{
A ggplot2 object (scatterplot)
}
\description{
This function enables a quick visualization (scatterplot) of the LOF scores alongside the data. This is only implemented for 2 variables.
}
\author{
Emanuel Sommer
}
