% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_dqtest.R
\name{perform_dqtest}
\alias{perform_dqtest}
\title{Univariate data quality testing.}
\usage{
perform_dqtest(
  x,
  var_name = "",
  cat_thres = 30,
  categories = NULL,
  range_min = NULL,
  range_max = NULL,
  tz = NULL,
  exclude_values = NULL,
  exclude_smaller_than = NULL,
  exclude_greater_than = NULL,
  plot_col1 = "darkviolet",
  plot_col2 = NULL,
  bin_num = 100,
  alpha = 0.8,
  rel = FALSE,
  orig_proportions = TRUE
)
}
\arguments{
\item{x}{A single vector (the allowed classes are: numeric, integer, factor, logical, character, POSIXct/ POSIXlt for datetimes and Date for dates.)}

\item{var_name}{The variable name to be used in diagnostic plots. Must be given as character.}

\item{cat_thres}{A non negative integer value to set the number of unique values a numeric or integer field needs to have to be classified as continuous. (default is set to 30)}

\item{categories}{A character vector of categories. For categorical input the categories check will be based on this vector.}

\item{range_min}{For continuous input data a numeric, for Date input a Date and for datetime input a POSIXct value. The value is themn used in the range check as the lower range bound.}

\item{range_max}{For continuous input data a numeric, for Date input a Date and for datetime input a POSIXct value. The value is themn used in the range check as the upper range bound.}

\item{tz}{If the output should be in another timezone than the input data, one can set the timezone here by supplying a character vector with a valid timezone shortcut.}

\item{exclude_values}{For continuous input data a numeric, for categorical input a character, for Date input a Date and for datetime input a POSIXct vector. The values of this vector will be excluded prior to the analysis.}

\item{exclude_smaller_than}{For continuous input data a numeric, for Date input a Date and for datetime input a POSIXct value. values strictly smaller than this value will be excluded prior to the analysis.}

\item{exclude_greater_than}{For continuous input data a numeric, for Date input a Date and for datetime input a POSIXct value. values strictly greater than this value will be excluded prior to the analysis.}

\item{plot_col1}{The primary plot color. Must be given as character.}

\item{plot_col2}{The secondary plot color. For example for range bounds or unallowed categories. Must be given as character.}

\item{bin_num}{The number of bins of the histogram (for continuous, date and datetime input). Must be given as a positive integer value.}

\item{alpha}{A numeric value between 0 and 1 to adjust the transparancy of the plots.}

\item{rel}{A logical value indicating whether the count or the density should be shown on the y axis of the diagnostic plots. (default: count)}

\item{orig_proportions}{A logical value indicating whether for categorical input in the relative barplot the length before exclutions should be used for the calculation of the proportions. If TRUE the original length will be used.}
}
\value{
A list. More details above.
}
\description{
The function perform_dqtest supplies the user with a list that contains the most important information to  perform  a  univariate data quality test.
}
\details{
\itemize{
   \item{The elements of the list vary depending on the input except for the element \code{pre_scheduled_return} which indicates whether all the values of x are missing values (-> "Yes, all NA."), or whether all values of x were excluded prior to the analysis (-> "Yes, all not missing excluded.") else the element contains "No.".}
   \item{The function first of all classifies the input vector with the help of the \code{\link{classify_vector}} function and further analysis is performed depending on this classification. One can access the classification via the \code{classification} element of the list.}
   \item{The class type of the initial vector (like numeric or POSIXct) can be accessed via the \code{class_type} element.}
   \item{The original length of the input vector can be accessed via the \code{initial_length} element.}
   \item{The absolute and relative amount of NAs (missing values) can be accessed via the \code{abs_na} and \code{rel_na} elements.}
}

The other elements of the list are then dependend on the initial classification:
\describe{
   \item{\strong{continuous}}{\itemize{
   \item{The \code{exclusions} element either contains the character value "No exclusions." or a list with the relative and absolute amount of exclusions.}
   \item{The \code{stat_summary} element contains a list with the mean, standard deviation, skewness and the seq(0,1,0.1) quantiles (11).}
   \item{The \code{range_check} element contains a character value the tells the user about the result of the range check.}
   \item{The \code{boxplot} and \code{hist} return two ggplot objects for diagnostic purposes.}}}
   \item{\strong{categorical}}{\itemize{
   \item{The \code{exclusions} element either contains the character value "No exclusions." or a list with the relative and absolute amount of exclusions.}
   \item{The \code{stat_summary} element contains a tibble with the rows representing all unique categories and with the columns showing the absolute and relative appearances ordered descending by absolute appearances.}
   \item{The \code{cat_check} element contains a character value the tells the user about the result of the categories check (whether all actual categories are included in the supplied ones).}
   \item{The \code{barplot} returns a ggplot object for diagnostic purposes.}}}
   \item{\strong{date}}{\itemize{
   \item{The \code{exclusions} element either contains the character value "No exclusions." or a list with the relative and absolute amount of exclusions.}
   \item{The \code{stat_summary} element contains a list with the mean and the seq(0,1,0.1) quantiles (11).}
   \item{The \code{range_check} element contains a character value the tells the user about the result of the range check.}
   \item{The \code{hist_date},\code{hist_year}, \code{hist_month}, \code{hist_week},\code{hist_wday} and \code{hist_day} return ggplot objects for diagnostic purposes.}}}
   \item{\strong{datetime}}{\itemize{
   \item{The \code{exclusions} element either contains the character value "No exclusions." or a list with the relative and absolute amount of exclusions.}
   \item{The \code{stat_summary} element contains a list with the mean and the seq(0,1,0.1) quantiles (11).}
   \item{The \code{range_check} element contains a character value the tells the user about the result of the range check.}
   \item{The \code{hist_date},\code{hist_year}, \code{hist_month}, \code{hist_week},\code{hist_wday}, \code{hist_day}, \code{hist_hour}, \code{hist_minute} and \code{hist_second} return ggplot objects for diagnostic purposes.}}}
}
}
\seealso{
\code{\link{classify_vector}}, \code{\link{bar_cat}}, \code{\link{cont_boxplot}}, \code{\link{hist_cont}}, \code{\link{hist_date}}, \code{\link{hist_datetime}}
}
\author{
Emanuel Sommer
}
